Dimensions 1.2.0
Changes to the API used for DimensionsAddons

Added option to set DimensionsMode in Settings.yml which can be set to "PLAYER_HISTORY" and "LINKED_PORTALS"
PLAYER_HISTORY: The way Dimensions used to work, Portals will teleport players accordingly between worlds and not just between the portal world and the overworld.
LINKED_PORTALS: This mode is also using history to try and keep players from not losing their bases by teleporting to the wrong world. BUT portals now also get linked. The first player who uses a portal will link the portal with the portal of the destination (but the portal at the destination not) and other players who use the same portal will have the same destination.
Also added option UnlinkPortalsAfter to unlink portals if they have been unsused for specified time.

Added ReplacePortalRuins option in Settings.yml which checks for inactive portals or broken portals to replace instead of building a new one. This option can take 4 values:
- NONE (does not check for inactive portals or portal remains)
- FULL (Will only replace a full structure of a portal with filled corners)
- <float> (This option can take any number from 0.0 to 1.0 and is used to replace any portal-like structures that have a (<float>*100)% similarity to portal)*
- ANY (Does the exact same this as the previous option but has a preset value of 0.9 (90%))

Performance updates when checking for safe spot search